backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
supabase-py
